restructured code, and updated linting workflow to actually work
